+++ src/lib/src/shared/ThFilter.svelte
@@ -1,11 +1,11 @@
 <script lang="ts">
-    import type { TableHandlerInterface, FieldLike } from '$lib/src/shared'
+    import type { TableHandlerInterface } from '$lib/src/shared'
     import type { Check } from '$lib/src/client'
 
     type T = $$Generic<Row>
     type Props = {
         table  : TableHandlerInterface<T>,
-        field  : FieldLike<T>,
+        field  : string | ((row: T) => unknown),
         check ?: Check
     }
     let { table, field, check = undefined }: Props = $props()